+++ audionimbus/src/effect/reflection.rs
@@ -3,16 +3,10 @@ use super::equalizer::Equalizer;
 use super::EffectError;
 use crate::audio_buffer::{AudioBuffer, Sample};
 use crate::audio_settings::AudioSettings;
-use crate::callback::{CallbackInformation, ProgressCallback};
 use crate::context::Context;
-use crate::device::open_cl::OpenClDevice;
-use crate::device::radeon_rays::RadeonRaysDevice;
 use crate::device::true_audio_next::TrueAudioNextDevice;
 use crate::error::{to_option_error, SteamAudioError};
 use crate::ffi_wrapper::FFIWrapper;
-use crate::geometry::{Scene, SceneParams};
-use crate::probe::ProbeBatch;
-use crate::simulation::BakedDataIdentifier;
 use crate::{ChannelPointers, ChannelRequirement};
 use std::marker::PhantomData;
 
@@ -141,7 +135,9 @@ impl ReflectionEffectType for TrueAudioNext {
 #[cfg(doc)]
 use super::AmbisonicsDecodeEffect;
 #[cfg(doc)]
-use crate::simulation::{BakedDataVariation, SimulationOutputs, Simulator, Source};
+use crate::geometry::Scene;
+#[cfg(doc)]
+use crate::simulation::{SimulationOutputs, Simulator, Source};
 
 /// Applies the result of physics-based reflections simulation to an audio buffer.
 ///
@@ -826,158 +822,6 @@ impl<T: ReflectionEffectType> ReflectionEffectParams<T> {
         };
 
         FFIWrapper::new(reflection_effect_params)
-    }
-}
-
-/// Bakes a single layer of reflections data in a probe batch.
-///
-/// Only one bake can be in progress at any point in time.
-pub fn bake_reflections(
-    context: &Context,
-    reflections_bake_params: ReflectionsBakeParams,
-    progress_callback: Option<CallbackInformation<ProgressCallback>>,
-) {
-    let (callback, user_data) = if let Some(callback_information) = progress_callback {
-        (
-            Some(callback_information.callback),
-            callback_information.user_data,
-        )
-    } else {
-        (None, std::ptr::null_mut())
-    };
-
-    unsafe {
-        audionimbus_sys::iplReflectionsBakerBake(
-            context.raw_ptr(),
-            &mut audionimbus_sys::IPLReflectionsBakeParams::from(reflections_bake_params),
-            callback,
-            user_data,
-        );
-    }
-}
-
-/// Cancels any running bakes of pathing data.
-pub fn cancel_bake_reflections(context: &Context) {
-    unsafe { audionimbus_sys::iplReflectionsBakerCancelBake(context.raw_ptr()) }
-}
-
-/// Parameters used to control how reflections data is baked.
-#[derive(Debug, Copy, Clone)]
-pub struct ReflectionsBakeParams<'a> {
-    /// The scene in which the probes exist.
-    pub scene: &'a Scene,
-
-    /// A probe batch containing the probes at which reflections data should be baked.
-    pub probe_batch: &'a ProbeBatch,
-
-    /// The scene parameters.
-    pub scene_params: SceneParams<'a>,
-
-    /// An identifier for the data layer that should be baked.
-    /// The identifier determines what data is simulated and stored at each probe.
-    /// If the probe batch already contains data with this identifier, it will be overwritten.
-    pub identifier: &'a BakedDataIdentifier,
-
-    /// The types of data to save for each probe.
-    pub bake_flags: ReflectionsBakeFlags,
-
-    /// The number of rays to trace from each listener position when baking.
-    /// Increasing this number results in improved accuracy, at the cost of increased bake times.
-    pub num_rays: u32,
-
-    /// The number of directions to consider when generating diffusely-reflected rays when baking.
-    /// Increasing this number results in slightly improved accuracy of diffuse reflections.
-    pub num_diffuse_samples: u32,
-
-    /// The number of times each ray is reflected off of solid geometry.
-    /// Increasing this number results in longer reverb tails and improved accuracy, at the cost of increased bake times.
-    pub num_bounces: u32,
-
-    /// The length (in seconds) of the impulse responses to simulate.
-    /// Increasing this number allows the baked data to represent longer reverb tails (and hence larger spaces), at the cost of increased memory usage while baking.
-    pub simulated_duration: f32,
-
-    /// The length (in seconds) of the impulse responses to save at each probe.
-    /// Increasing this number allows the baked data to represent longer reverb tails (and hence larger spaces), at the cost of increased disk space usage and memory usage at run-time.
-    ///
-    /// It may be useful to set [`Self::saved_duration`] to be less than [`Self::simulated_duration`], especially if you plan to use hybrid reverb for rendering baked reflections.
-    /// This way, the parametric reverb data is estimated using a longer IR, resulting in more accurate estimation, but only the early part of the IR can be saved for subsequent rendering.
-    pub saved_duration: f32,
-
-    /// Ambisonic order of the baked IRs.
-    pub order: u32,
-
-    /// Number of threads to use for baking.
-    pub num_threads: u32,
-
-    /// When calculating how much sound energy reaches a surface directly from a source, any source that is closer than [`Self::irradiance_min_distance`] to the surface is assumed to be at a distance of [`Self::irradiance_min_distance`], for the purposes of energy calculations.
-    pub irradiance_min_distance: f32,
-
-    /// If using Radeon Rays or if [`Self::identifier`] uses [`BakedDataVariation::StaticListener`], this is the number of probes for which data is baked simultaneously.
-    pub bake_batch_size: u32,
-}
-
-impl From<ReflectionsBakeParams<'_>> for audionimbus_sys::IPLReflectionsBakeParams {
-    fn from(params: ReflectionsBakeParams) -> Self {
-        let mut ray_batch_size = 0;
-        let mut open_cl_device = &OpenClDevice::null();
-        let mut radeon_rays_device = &RadeonRaysDevice::null();
-        let scene_type = match params.scene_params {
-            SceneParams::Default => audionimbus_sys::IPLSceneType::IPL_SCENETYPE_DEFAULT,
-            SceneParams::Embree => audionimbus_sys::IPLSceneType::IPL_SCENETYPE_EMBREE,
-            SceneParams::RadeonRays {
-                open_cl_device: ocl_device,
-                radeon_rays_device: rr_device,
-            } => {
-                open_cl_device = ocl_device;
-                radeon_rays_device = rr_device;
-                audionimbus_sys::IPLSceneType::IPL_SCENETYPE_RADEONRAYS
-            }
-            SceneParams::Custom {
-                ray_batch_size: rb_size,
-            } => {
-                ray_batch_size = rb_size;
-                audionimbus_sys::IPLSceneType::IPL_SCENETYPE_CUSTOM
-            }
-        };
-
-        Self {
-            scene: params.scene.raw_ptr(),
-            probeBatch: params.probe_batch.raw_ptr(),
-            sceneType: scene_type,
-            identifier: (*params.identifier).into(),
-            bakeFlags: params.bake_flags.into(),
-            numRays: params.num_rays as i32,
-            numDiffuseSamples: params.num_diffuse_samples as i32,
-            numBounces: params.num_bounces as i32,
-            simulatedDuration: params.simulated_duration,
-            savedDuration: params.saved_duration,
-            order: params.order as i32,
-            numThreads: params.num_threads as i32,
-            rayBatchSize: ray_batch_size as i32,
-            irradianceMinDistance: params.irradiance_min_distance,
-            bakeBatchSize: params.bake_batch_size as i32,
-            openCLDevice: open_cl_device.raw_ptr(),
-            radeonRaysDevice: radeon_rays_device.raw_ptr(),
-        }
-    }
-}
-
-bitflags::bitflags! {
-    /// Flags for specifying what types of reflections data to bake.
-    #[derive(Copy, Clone, Debug)]
-    pub struct ReflectionsBakeFlags: u32 {
-        /// Bake impulse responses for [`ReflectionEffectSettings::Convolution`], [`ReflectionEffectSettings::Hybrid`], or [`ReflectionEffectSettings::TrueAudioNext`].
-        const BAKE_CONVOLUTION = 1 << 0;
-
-        /// Bake parametric reverb for [`ReflectionEffectSettings::Parametric`] or [`ReflectionEffectSettings::Hybrid`].
-        const BAKE_PARAMETRIC = 1 << 1;
-    }
-}
-
-impl From<ReflectionsBakeFlags> for audionimbus_sys::IPLReflectionsBakeFlags {
-    fn from(reflections_bake_flags: ReflectionsBakeFlags) -> Self {
-        Self(reflections_bake_flags.bits() as _)
     }
 }